feat: added empty ubos handling & refactor
