+++ apps/backoffice-v2/src/lib/blocks/hooks/useUbosUserProvidedBlock/types.ts
@@ -1,0 +1,8 @@
+export interface IUBOSUserProvided {
+  name: string;
+  nationality: string;
+  identityNumber: string;
+  percentageOfOwnership: number;
+  email: string;
+  address: string;
+}
